Normalize path to match platform
The warning about the renamed target not existing in the destination
directory was incorrectly being shown on Windows if the dest path used
forward slashes. This was due to a comparison against a Windows-style
path with back slashes. To correctly handle this, we just normalize the
destination path before comparing it to a computed path.